requirements.txt
discord.py
beautifulsoup4
pymongo[srv]
Pillow
requests
aiohttp
pydeepl
unidecode
